实现了GetAllGoal

+++ surakarta_/surakarta/surakarta_board.h
@@ -22,8 +22,8 @@ public:
     SurakartaBoard(unsigned int );
     ~SurakartaBoard();
     void reset();
-    int cell_width=50;             //格子宽度
-    int  piece_radius=cell_width*3/7;// r棋子的半径
+    double cell_width=50;             //格子宽度
+    double piece_radius=cell_width*3/7;// r棋子的半径
     int selectId;              // 被选中的棋子
     bool isBlackTurn;              // 是否轮到黑方走
     SurakartaPiece piece[24];
